+++ workflow_code/run_workflow.py
@@ -287,7 +287,7 @@ manifest {{
     mainScript = 'main.nf'
     defaultBranch = 'main'
     nextflowVersion = '>=24.04.4'
-    version = '1.0.5'
+    version = '1.0.6'
 }}
 
     """